refactor(registry): split client module
